requirements.txt
Django==1.8.5
django-vanilla-views==1.0.4
djangorestframework==3.2.4
jsonfield==1.0.3
wheel==0.24.0
